add new search, featuring entity and member search
This commit adds a revamped search function for the scaladoc tool. It
also contains a number of small fixes for HTML-layout and JavaScript
issues.

The search is implemented by enhancing the scheduler and using
JavaScript promises. List of changes/additions:

* Revamped search functionality
    - Search members as well as entities
    - Preserve keyboard navigation
    - Scroll to selected entity if outside of viewport
    - Non-blocking, cancelable
* Display of library name (top left)
* Refactored scheduler
* Cleanup of HTML layout
    - Light left pane
    - Better mobile layout, no need for dynamic offsets
    - Remove unused element classes

Known issues
------------
Some classes that have a lot of implicit conversions (i.e. Array) will
show multiple members with the same signature. This is an internal issue
with scaladoc. This is further explained by @VladUreche in SI-9620. Once
it has been solved, this issue will be gone.